+++ data/src/rsbudget/data/api/bo/MonetaryValue.java
@@ -3,6 +3,8 @@
  */
 package rsbudget.data.api.bo;
 
+import java.math.BigDecimal;
+
 /**
  * Defines an interface returning monetary values.
  * @author Ralph
@@ -14,11 +16,11 @@ public interface MonetaryValue {
 	 * Returns the amount.
 	 * @return the amount
 	 */
-	public float getAmount();
+	public BigDecimal getAmount();
 	
 	/**
 	 * Sets the amount.
 	 * @param amount the value
 	 */
-	public void setAmount(float amount);
+	public void setAmount(BigDecimal amount);
 }
