+++ data/src/rsbudget/data/impl/bo/PlannedTransactionBO.java
@@ -3,6 +3,8 @@
  */
 package rsbudget.data.impl.bo;
 
+import java.math.BigDecimal;
+
 import rsbudget.data.api.bo.Account;
 import rsbudget.data.api.bo.Budget;
 import rsbudget.data.api.bo.Category;
@@ -42,7 +44,7 @@ public class PlannedTransactionBO extends AbstractRsBudgetDbBO<PlannedTransactio
 	 * {@inheritDoc}
 	 */
 	@Override
-	public float getAmount() {
+	public BigDecimal getAmount() {
 		return getTransferObject().getAmount();
 	}
 
@@ -50,8 +52,8 @@ public class PlannedTransactionBO extends AbstractRsBudgetDbBO<PlannedTransactio
 	 * {@inheritDoc}
 	 */
 	@Override
-	public void setAmount(float amount) {
-		float oldValue = getAmount();
+	public void setAmount(BigDecimal amount) {
+		BigDecimal oldValue = getAmount();
 		getTransferObject().setAmount(amount);
 		firePropertyChange(PROPERTY_AMOUNT, oldValue, amount);
 	}
